# lib/api-zod/src/generated/types/csvPreview.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI-powered CSV Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { CsvPreviewRowsItem } from './csvPreviewRowsItem';

export interface CsvPreview {
  headers: string[];
  rows: CsvPreviewRowsItem[];
  totalRows: number;
  fileName: string;
}