# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI-powered CSV Importer API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
  /** @nullable */
  details?: string | null;
}