# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI-powered CSV Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { CrmRecord } from './crmRecord';
import type { SkippedRecord } from './skippedRecord';

export interface ImportResult {
  records: CrmRecord[];
  skipped: SkippedRecord[];
  totalImported: number;
  totalSkipped: number;
  totalProcessed: number;
}