# lib/api-zod/src/generated/types/skippedRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowEasy AI-powered CSV Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { SkippedRecordRawData } from './skippedRecordRawData';

export interface SkippedRecord {
  rowIndex: number;
  reason: string;
  rawData: SkippedRecordRawData;
}